IDE-3022 revert isBundleOnlyProject and support projects under lr-ws/wars
