+++ tools/plugins/com.liferay.ide.server.ui/src/com/liferay/ide/server/ui/portal/BundleModuleArtifactAdapterDelegate.java
@@ -15,6 +15,7 @@
 package com.liferay.ide.server.ui.portal;
 
 import com.liferay.ide.core.util.CoreUtil;
+import com.liferay.ide.project.core.util.ProjectUtil;
 
 import org.eclipse.core.resources.IProject;
 import org.eclipse.core.runtime.IAdaptable;
@@ -78,7 +79,7 @@ public class BundleModuleArtifactAdapterDelegate extends ModuleArtifactAdapterDe
 
        if( project != null )
        {
-            if( CoreUtil.isBundleOnlyProject( project ) )
+            if( CoreUtil.isBundleOnlyProject( project ) || ProjectUtil.isFacetedGradleBundleProject( project ) )
             {
                 return new WebResource( getModule( project ), project.getProjectRelativePath() );
             }
